@@ -24,17 +24,19 @@
   "devDependencies": {
     "@faker-js/faker": "^8.0.0",
     "@floating-ui/dom": "^1.5.1",
+    "@socket.io/component-emitter": "^3.1.0",
     "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/kit": "^2.0.6",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/enhanced-img": "^0.1.8",
+    "@sveltejs/kit": "^2.5.0",
+    "@sveltejs/vite-plugin-svelte": "^3.0.2",
     "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/svelte": "^4.0.3",
     "@types/dom-to-image": "^2.6.4",
     "@types/justified-layout": "^4.1.0",
     "@types/lodash-es": "^4.17.6",
     "@types/luxon": "^3.2.0",
-    "@typescript-eslint/eslint-plugin": "^6.4.1",
-    "@typescript-eslint/parser": "^6.4.1",
+    "@typescript-eslint/eslint-plugin": "^7.0.0",
+    "@typescript-eslint/parser": "^7.0.0",
     "@vitest/coverage-v8": "^1.0.4",
     "autoprefixer": "^10.4.13",
     "eslint": "^8.34.0",
@@ -48,9 +50,8 @@
     "prettier-plugin-organize-imports": "^3.2.4",
     "prettier-plugin-svelte": "^3.1.2",
     "rollup-plugin-visualizer": "^5.12.0",
-    "svelte": "^4.0.5",
-    "svelte-check": "^3.4.3",
-    "svelte-preprocess": "^5.0.3",
+    "svelte": "^4.2.11",
+    "svelte-check": "^3.6.4",
     "tailwindcss": "^3.2.7",
     "tslib": "^2.5.0",
     "typescript": "^5.3.3",
@@ -73,7 +74,7 @@
     "luxon": "^3.2.1",
     "socket.io-client": "^4.6.1",
     "svelte-local-storage-store": "^0.6.0",
-    "svelte-maplibre": "^0.7.0",
+    "svelte-maplibre": "^0.8.0",
     "thumbhash": "^0.1.1"
   }
 }
