@@ -15,7 +15,6 @@ class ThumbnailImage extends StatelessWidget {
   final int totalAssets;
   final bool showStorageIndicator;
   final bool showStack;
-  final bool useGrayBoxPlaceholder;
   final bool isSelected;
   final bool multiselectEnabled;
   final Function? onSelect;
@@ -30,7 +29,6 @@ class ThumbnailImage extends StatelessWidget {
     required this.totalAssets,
     this.showStorageIndicator = true,
     this.showStack = false,
-    this.useGrayBoxPlaceholder = false,
     this.isSelected = false,
     this.multiselectEnabled = false,
     this.onDeselect,
@@ -138,6 +136,8 @@ class ThumbnailImage extends StatelessWidget {
               : asset.id + heroOffset,
           child: ImmichImage.thumbnail(
             asset,
+            height: 300,
+            width: 300,
           ),
         ),
       );
