@@ -254,231 +254,27 @@ class SearchApi {
     return null;
   }
 
-  /// Performs an HTTP 'GET /search/metadata' operation and returns the [Response].
+  /// Performs an HTTP 'POST /search/metadata' operation and returns the [Response].
   /// Parameters:
   ///
-  /// * [String] checksum:
-  ///
-  /// * [String] city:
-  ///
-  /// * [String] country:
-  ///
-  /// * [DateTime] createdAfter:
-  ///
-  /// * [DateTime] createdBefore:
-  ///
-  /// * [String] deviceAssetId:
-  ///
-  /// * [String] deviceId:
-  ///
-  /// * [String] encodedVideoPath:
-  ///
-  /// * [String] id:
-  ///
-  /// * [bool] isArchived:
-  ///
-  /// * [bool] isEncoded:
-  ///
-  /// * [bool] isExternal:
-  ///
-  /// * [bool] isFavorite:
-  ///
-  /// * [bool] isMotion:
-  ///
-  /// * [bool] isOffline:
-  ///
-  /// * [bool] isReadOnly:
-  ///
-  /// * [bool] isVisible:
-  ///
-  /// * [String] lensModel:
-  ///
-  /// * [String] libraryId:
-  ///
-  /// * [String] make:
-  ///
-  /// * [String] model:
-  ///
-  /// * [AssetOrder] order:
-  ///
-  /// * [String] originalFileName:
-  ///
-  /// * [String] originalPath:
-  ///
-  /// * [num] page:
-  ///
-  /// * [String] resizePath:
-  ///
-  /// * [num] size:
-  ///
-  /// * [String] state:
-  ///
-  /// * [DateTime] takenAfter:
-  ///
-  /// * [DateTime] takenBefore:
-  ///
-  /// * [DateTime] trashedAfter:
-  ///
-  /// * [DateTime] trashedBefore:
-  ///
-  /// * [AssetTypeEnum] type:
-  ///
-  /// * [DateTime] updatedAfter:
-  ///
-  /// * [DateTime] updatedBefore:
-  ///
-  /// * [String] webpPath:
-  ///
-  /// * [bool] withArchived:
-  ///
-  /// * [bool] withDeleted:
-  ///
-  /// * [bool] withExif:
-  ///
-  /// * [bool] withPeople:
-  ///
-  /// * [bool] withStacked:
-  Future<Response> searchMetadataWithHttpInfo({ String? checksum, String? city, String? country, DateTime? createdAfter, DateTime? createdBefore, String? deviceAssetId, String? deviceId, String? encodedVideoPath, String? id, bool? isArchived, bool? isEncoded, bool? isExternal, bool? isFavorite, bool? isMotion, bool? isOffline, bool? isReadOnly, bool? isVisible, String? lensModel, String? libraryId, String? make, String? model, AssetOrder? order, String? originalFileName, String? originalPath, num? page, String? resizePath, num? size, String? state, DateTime? takenAfter, DateTime? takenBefore, DateTime? trashedAfter, DateTime? trashedBefore, AssetTypeEnum? type, DateTime? updatedAfter, DateTime? updatedBefore, String? webpPath, bool? withArchived, bool? withDeleted, bool? withExif, bool? withPeople, bool? withStacked, }) async {
+  /// * [MetadataSearchDto] metadataSearchDto (required):
+  Future<Response> searchMetadataWithHttpInfo(MetadataSearchDto metadataSearchDto,) async {
     // ignore: prefer_const_declarations
     final path = r'/search/metadata';
 
     // ignore: prefer_final_locals
-    Object? postBody;
+    Object? postBody = metadataSearchDto;
 
     final queryParams = <QueryParam>[];
     final headerParams = <String, String>{};
     final formParams = <String, String>{};
 
-    if (checksum != null) {
-      queryParams.addAll(_queryParams('', 'checksum', checksum));
-    }
-    if (city != null) {
-      queryParams.addAll(_queryParams('', 'city', city));
-    }
-    if (country != null) {
-      queryParams.addAll(_queryParams('', 'country', country));
-    }
-    if (createdAfter != null) {
-      queryParams.addAll(_queryParams('', 'createdAfter', createdAfter));
-    }
-    if (createdBefore != null) {
-      queryParams.addAll(_queryParams('', 'createdBefore', createdBefore));
-    }
-    if (deviceAssetId != null) {
-      queryParams.addAll(_queryParams('', 'deviceAssetId', deviceAssetId));
-    }
-    if (deviceId != null) {
-      queryParams.addAll(_queryParams('', 'deviceId', deviceId));
-    }
-    if (encodedVideoPath != null) {
-      queryParams.addAll(_queryParams('', 'encodedVideoPath', encodedVideoPath));
-    }
-    if (id != null) {
-      queryParams.addAll(_queryParams('', 'id', id));
-    }
-    if (isArchived != null) {
-      queryParams.addAll(_queryParams('', 'isArchived', isArchived));
-    }
-    if (isEncoded != null) {
-      queryParams.addAll(_queryParams('', 'isEncoded', isEncoded));
-    }
-    if (isExternal != null) {
-      queryParams.addAll(_queryParams('', 'isExternal', isExternal));
-    }
-    if (isFavorite != null) {
-      queryParams.addAll(_queryParams('', 'isFavorite', isFavorite));
-    }
-    if (isMotion != null) {
-      queryParams.addAll(_queryParams('', 'isMotion', isMotion));
-    }
-    if (isOffline != null) {
-      queryParams.addAll(_queryParams('', 'isOffline', isOffline));
-    }
-    if (isReadOnly != null) {
-      queryParams.addAll(_queryParams('', 'isReadOnly', isReadOnly));
-    }
-    if (isVisible != null) {
-      queryParams.addAll(_queryParams('', 'isVisible', isVisible));
-    }
-    if (lensModel != null) {
-      queryParams.addAll(_queryParams('', 'lensModel', lensModel));
-    }
-    if (libraryId != null) {
-      queryParams.addAll(_queryParams('', 'libraryId', libraryId));
-    }
-    if (make != null) {
-      queryParams.addAll(_queryParams('', 'make', make));
-    }
-    if (model != null) {
-      queryParams.addAll(_queryParams('', 'model', model));
-    }
-    if (order != null) {
-      queryParams.addAll(_queryParams('', 'order', order));
-    }
-    if (originalFileName != null) {
-      queryParams.addAll(_queryParams('', 'originalFileName', originalFileName));
-    }
-    if (originalPath != null) {
-      queryParams.addAll(_queryParams('', 'originalPath', originalPath));
-    }
-    if (page != null) {
-      queryParams.addAll(_queryParams('', 'page', page));
-    }
-    if (resizePath != null) {
-      queryParams.addAll(_queryParams('', 'resizePath', resizePath));
-    }
-    if (size != null) {
-      queryParams.addAll(_queryParams('', 'size', size));
-    }
-    if (state != null) {
-      queryParams.addAll(_queryParams('', 'state', state));
-    }
-    if (takenAfter != null) {
-      queryParams.addAll(_queryParams('', 'takenAfter', takenAfter));
-    }
-    if (takenBefore != null) {
-      queryParams.addAll(_queryParams('', 'takenBefore', takenBefore));
-    }
-    if (trashedAfter != null) {
-      queryParams.addAll(_queryParams('', 'trashedAfter', trashedAfter));
-    }
-    if (trashedBefore != null) {
-      queryParams.addAll(_queryParams('', 'trashedBefore', trashedBefore));
-    }
-    if (type != null) {
-      queryParams.addAll(_queryParams('', 'type', type));
-    }
-    if (updatedAfter != null) {
-      queryParams.addAll(_queryParams('', 'updatedAfter', updatedAfter));
-    }
-    if (updatedBefore != null) {
-      queryParams.addAll(_queryParams('', 'updatedBefore', updatedBefore));
-    }
-    if (webpPath != null) {
-      queryParams.addAll(_queryParams('', 'webpPath', webpPath));
-    }
-    if (withArchived != null) {
-      queryParams.addAll(_queryParams('', 'withArchived', withArchived));
-    }
-    if (withDeleted != null) {
-      queryParams.addAll(_queryParams('', 'withDeleted', withDeleted));
-    }
-    if (withExif != null) {
-      queryParams.addAll(_queryParams('', 'withExif', withExif));
-    }
-    if (withPeople != null) {
-      queryParams.addAll(_queryParams('', 'withPeople', withPeople));
-    }
-    if (withStacked != null) {
-      queryParams.addAll(_queryParams('', 'withStacked', withStacked));
-    }
-
-    const contentTypes = <String>[];
+    const contentTypes = <String>['application/json'];
 
 
     return apiClient.invokeAPI(
       path,
-      'GET',
+      'POST',
       queryParams,
       postBody,
       headerParams,
@@ -489,89 +285,9 @@ class SearchApi {
 
   /// Parameters:
   ///
-  /// * [String] checksum:
-  ///
-  /// * [String] city:
-  ///
-  /// * [String] country:
-  ///
-  /// * [DateTime] createdAfter:
-  ///
-  /// * [DateTime] createdBefore:
-  ///
-  /// * [String] deviceAssetId:
-  ///
-  /// * [String] deviceId:
-  ///
-  /// * [String] encodedVideoPath:
-  ///
-  /// * [String] id:
-  ///
-  /// * [bool] isArchived:
-  ///
-  /// * [bool] isEncoded:
-  ///
-  /// * [bool] isExternal:
-  ///
-  /// * [bool] isFavorite:
-  ///
-  /// * [bool] isMotion:
-  ///
-  /// * [bool] isOffline:
-  ///
-  /// * [bool] isReadOnly:
-  ///
-  /// * [bool] isVisible:
-  ///
-  /// * [String] lensModel:
-  ///
-  /// * [String] libraryId:
-  ///
-  /// * [String] make:
-  ///
-  /// * [String] model:
-  ///
-  /// * [AssetOrder] order:
-  ///
-  /// * [String] originalFileName:
-  ///
-  /// * [String] originalPath:
-  ///
-  /// * [num] page:
-  ///
-  /// * [String] resizePath:
-  ///
-  /// * [num] size:
-  ///
-  /// * [String] state:
-  ///
-  /// * [DateTime] takenAfter:
-  ///
-  /// * [DateTime] takenBefore:
-  ///
-  /// * [DateTime] trashedAfter:
-  ///
-  /// * [DateTime] trashedBefore:
-  ///
-  /// * [AssetTypeEnum] type:
-  ///
-  /// * [DateTime] updatedAfter:
-  ///
-  /// * [DateTime] updatedBefore:
-  ///
-  /// * [String] webpPath:
-  ///
-  /// * [bool] withArchived:
-  ///
-  /// * [bool] withDeleted:
-  ///
-  /// * [bool] withExif:
-  ///
-  /// * [bool] withPeople:
-  ///
-  /// * [bool] withStacked:
-  Future<SearchResponseDto?> searchMetadata({ String? checksum, String? city, String? country, DateTime? createdAfter, DateTime? createdBefore, String? deviceAssetId, String? deviceId, String? encodedVideoPath, String? id, bool? isArchived, bool? isEncoded, bool? isExternal, bool? isFavorite, bool? isMotion, bool? isOffline, bool? isReadOnly, bool? isVisible, String? lensModel, String? libraryId, String? make, String? model, AssetOrder? order, String? originalFileName, String? originalPath, num? page, String? resizePath, num? size, String? state, DateTime? takenAfter, DateTime? takenBefore, DateTime? trashedAfter, DateTime? trashedBefore, AssetTypeEnum? type, DateTime? updatedAfter, DateTime? updatedBefore, String? webpPath, bool? withArchived, bool? withDeleted, bool? withExif, bool? withPeople, bool? withStacked, }) async {
-    final response = await searchMetadataWithHttpInfo( checksum: checksum, city: city, country: country, createdAfter: createdAfter, createdBefore: createdBefore, deviceAssetId: deviceAssetId, deviceId: deviceId, encodedVideoPath: encodedVideoPath, id: id, isArchived: isArchived, isEncoded: isEncoded, isExternal: isExternal, isFavorite: isFavorite, isMotion: isMotion, isOffline: isOffline, isReadOnly: isReadOnly, isVisible: isVisible, lensModel: lensModel, libraryId: libraryId, make: make, model: model, order: order, originalFileName: originalFileName, originalPath: originalPath, page: page, resizePath: resizePath, size: size, state: state, takenAfter: takenAfter, takenBefore: takenBefore, trashedAfter: trashedAfter, trashedBefore: trashedBefore, type: type, updatedAfter: updatedAfter, updatedBefore: updatedBefore, webpPath: webpPath, withArchived: withArchived, withDeleted: withDeleted, withExif: withExif, withPeople: withPeople, withStacked: withStacked, );
+  /// * [MetadataSearchDto] metadataSearchDto (required):
+  Future<SearchResponseDto?> searchMetadata(MetadataSearchDto metadataSearchDto,) async {
+    final response = await searchMetadataWithHttpInfo(metadataSearchDto,);
     if (response.statusCode >= HttpStatus.badRequest) {
       throw ApiException(response.statusCode, await _decodeBodyBytes(response));
     }
@@ -644,179 +360,27 @@ class SearchApi {
     return null;
   }
 
-  /// Performs an HTTP 'GET /search/smart' operation and returns the [Response].
+  /// Performs an HTTP 'POST /search/smart' operation and returns the [Response].
   /// Parameters:
   ///
-  /// * [String] query (required):
-  ///
-  /// * [String] city:
-  ///
-  /// * [String] country:
-  ///
-  /// * [DateTime] createdAfter:
-  ///
-  /// * [DateTime] createdBefore:
-  ///
-  /// * [String] deviceId:
-  ///
-  /// * [bool] isArchived:
-  ///
-  /// * [bool] isEncoded:
-  ///
-  /// * [bool] isExternal:
-  ///
-  /// * [bool] isFavorite:
-  ///
-  /// * [bool] isMotion:
-  ///
-  /// * [bool] isOffline:
-  ///
-  /// * [bool] isReadOnly:
-  ///
-  /// * [bool] isVisible:
-  ///
-  /// * [String] lensModel:
-  ///
-  /// * [String] libraryId:
-  ///
-  /// * [String] make:
-  ///
-  /// * [String] model:
-  ///
-  /// * [num] page:
-  ///
-  /// * [num] size:
-  ///
-  /// * [String] state:
-  ///
-  /// * [DateTime] takenAfter:
-  ///
-  /// * [DateTime] takenBefore:
-  ///
-  /// * [DateTime] trashedAfter:
-  ///
-  /// * [DateTime] trashedBefore:
-  ///
-  /// * [AssetTypeEnum] type:
-  ///
-  /// * [DateTime] updatedAfter:
-  ///
-  /// * [DateTime] updatedBefore:
-  ///
-  /// * [bool] withArchived:
-  ///
-  /// * [bool] withDeleted:
-  ///
-  /// * [bool] withExif:
-  Future<Response> searchSmartWithHttpInfo(String query, { String? city, String? country, DateTime? createdAfter, DateTime? createdBefore, String? deviceId, bool? isArchived, bool? isEncoded, bool? isExternal, bool? isFavorite, bool? isMotion, bool? isOffline, bool? isReadOnly, bool? isVisible, String? lensModel, String? libraryId, String? make, String? model, num? page, num? size, String? state, DateTime? takenAfter, DateTime? takenBefore, DateTime? trashedAfter, DateTime? trashedBefore, AssetTypeEnum? type, DateTime? updatedAfter, DateTime? updatedBefore, bool? withArchived, bool? withDeleted, bool? withExif, }) async {
+  /// * [SmartSearchDto] smartSearchDto (required):
+  Future<Response> searchSmartWithHttpInfo(SmartSearchDto smartSearchDto,) async {
     // ignore: prefer_const_declarations
     final path = r'/search/smart';
 
     // ignore: prefer_final_locals
-    Object? postBody;
+    Object? postBody = smartSearchDto;
 
     final queryParams = <QueryParam>[];
     final headerParams = <String, String>{};
     final formParams = <String, String>{};
 
-    if (city != null) {
-      queryParams.addAll(_queryParams('', 'city', city));
-    }
-    if (country != null) {
-      queryParams.addAll(_queryParams('', 'country', country));
-    }
-    if (createdAfter != null) {
-      queryParams.addAll(_queryParams('', 'createdAfter', createdAfter));
-    }
-    if (createdBefore != null) {
-      queryParams.addAll(_queryParams('', 'createdBefore', createdBefore));
-    }
-    if (deviceId != null) {
-      queryParams.addAll(_queryParams('', 'deviceId', deviceId));
-    }
-    if (isArchived != null) {
-      queryParams.addAll(_queryParams('', 'isArchived', isArchived));
-    }
-    if (isEncoded != null) {
-      queryParams.addAll(_queryParams('', 'isEncoded', isEncoded));
-    }
-    if (isExternal != null) {
-      queryParams.addAll(_queryParams('', 'isExternal', isExternal));
-    }
-    if (isFavorite != null) {
-      queryParams.addAll(_queryParams('', 'isFavorite', isFavorite));
-    }
-    if (isMotion != null) {
-      queryParams.addAll(_queryParams('', 'isMotion', isMotion));
-    }
-    if (isOffline != null) {
-      queryParams.addAll(_queryParams('', 'isOffline', isOffline));
-    }
-    if (isReadOnly != null) {
-      queryParams.addAll(_queryParams('', 'isReadOnly', isReadOnly));
-    }
-    if (isVisible != null) {
-      queryParams.addAll(_queryParams('', 'isVisible', isVisible));
-    }
-    if (lensModel != null) {
-      queryParams.addAll(_queryParams('', 'lensModel', lensModel));
-    }
-    if (libraryId != null) {
-      queryParams.addAll(_queryParams('', 'libraryId', libraryId));
-    }
-    if (make != null) {
-      queryParams.addAll(_queryParams('', 'make', make));
-    }
-    if (model != null) {
-      queryParams.addAll(_queryParams('', 'model', model));
-    }
-    if (page != null) {
-      queryParams.addAll(_queryParams('', 'page', page));
-    }
-      queryParams.addAll(_queryParams('', 'query', query));
-    if (size != null) {
-      queryParams.addAll(_queryParams('', 'size', size));
-    }
-    if (state != null) {
-      queryParams.addAll(_queryParams('', 'state', state));
-    }
-    if (takenAfter != null) {
-      queryParams.addAll(_queryParams('', 'takenAfter', takenAfter));
-    }
-    if (takenBefore != null) {
-      queryParams.addAll(_queryParams('', 'takenBefore', takenBefore));
-    }
-    if (trashedAfter != null) {
-      queryParams.addAll(_queryParams('', 'trashedAfter', trashedAfter));
-    }
-    if (trashedBefore != null) {
-      queryParams.addAll(_queryParams('', 'trashedBefore', trashedBefore));
-    }
-    if (type != null) {
-      queryParams.addAll(_queryParams('', 'type', type));
-    }
-    if (updatedAfter != null) {
-      queryParams.addAll(_queryParams('', 'updatedAfter', updatedAfter));
-    }
-    if (updatedBefore != null) {
-      queryParams.addAll(_queryParams('', 'updatedBefore', updatedBefore));
-    }
-    if (withArchived != null) {
-      queryParams.addAll(_queryParams('', 'withArchived', withArchived));
-    }
-    if (withDeleted != null) {
-      queryParams.addAll(_queryParams('', 'withDeleted', withDeleted));
-    }
-    if (withExif != null) {
-      queryParams.addAll(_queryParams('', 'withExif', withExif));
-    }
-
-    const contentTypes = <String>[];
+    const contentTypes = <String>['application/json'];
 
 
     return apiClient.invokeAPI(
       path,
-      'GET',
+      'POST',
       queryParams,
       postBody,
       headerParams,
@@ -827,69 +391,9 @@ class SearchApi {
 
   /// Parameters:
   ///
-  /// * [String] query (required):
-  ///
-  /// * [String] city:
-  ///
-  /// * [String] country:
-  ///
-  /// * [DateTime] createdAfter:
-  ///
-  /// * [DateTime] createdBefore:
-  ///
-  /// * [String] deviceId:
-  ///
-  /// * [bool] isArchived:
-  ///
-  /// * [bool] isEncoded:
-  ///
-  /// * [bool] isExternal:
-  ///
-  /// * [bool] isFavorite:
-  ///
-  /// * [bool] isMotion:
-  ///
-  /// * [bool] isOffline:
-  ///
-  /// * [bool] isReadOnly:
-  ///
-  /// * [bool] isVisible:
-  ///
-  /// * [String] lensModel:
-  ///
-  /// * [String] libraryId:
-  ///
-  /// * [String] make:
-  ///
-  /// * [String] model:
-  ///
-  /// * [num] page:
-  ///
-  /// * [num] size:
-  ///
-  /// * [String] state:
-  ///
-  /// * [DateTime] takenAfter:
-  ///
-  /// * [DateTime] takenBefore:
-  ///
-  /// * [DateTime] trashedAfter:
-  ///
-  /// * [DateTime] trashedBefore:
-  ///
-  /// * [AssetTypeEnum] type:
-  ///
-  /// * [DateTime] updatedAfter:
-  ///
-  /// * [DateTime] updatedBefore:
-  ///
-  /// * [bool] withArchived:
-  ///
-  /// * [bool] withDeleted:
-  ///
-  /// * [bool] withExif:
-  Future<SearchResponseDto?> searchSmart(String query, { String? city, String? country, DateTime? createdAfter, DateTime? createdBefore, String? deviceId, bool? isArchived, bool? isEncoded, bool? isExternal, bool? isFavorite, bool? isMotion, bool? isOffline, bool? isReadOnly, bool? isVisible, String? lensModel, String? libraryId, String? make, String? model, num? page, num? size, String? state, DateTime? takenAfter, DateTime? takenBefore, DateTime? trashedAfter, DateTime? trashedBefore, AssetTypeEnum? type, DateTime? updatedAfter, DateTime? updatedBefore, bool? withArchived, bool? withDeleted, bool? withExif, }) async {
-    final response = await searchSmartWithHttpInfo(query,  city: city, country: country, createdAfter: createdAfter, createdBefore: createdBefore, deviceId: deviceId, isArchived: isArchived, isEncoded: isEncoded, isExternal: isExternal, isFavorite: isFavorite, isMotion: isMotion, isOffline: isOffline, isReadOnly: isReadOnly, isVisible: isVisible, lensModel: lensModel, libraryId: libraryId, make: make, model: model, page: page, size: size, state: state, takenAfter: takenAfter, takenBefore: takenBefore, trashedAfter: trashedAfter, trashedBefore: trashedBefore, type: type, updatedAfter: updatedAfter, updatedBefore: updatedBefore, withArchived: withArchived, withDeleted: withDeleted, withExif: withExif, );
+  /// * [SmartSearchDto] smartSearchDto (required):
+  Future<SearchResponseDto?> searchSmart(SmartSearchDto smartSearchDto,) async {
+    final response = await searchSmartWithHttpInfo(smartSearchDto,);
     if (response.statusCode >= HttpStatus.badRequest) {
       throw ApiException(response.statusCode, await _decodeBodyBytes(response));
     }
