@@ -2333,6 +2333,273 @@ export interface MergePersonDto {
 /**
  * 
  * @export
+ * @interface MetadataSearchDto
+ */
+export interface MetadataSearchDto {
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'checksum'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'city'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'country'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'createdAfter'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'createdBefore'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'deviceAssetId'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'deviceId'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'encodedVideoPath'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'id'?: string;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'isArchived'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'isEncoded'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'isExternal'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'isFavorite'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'isMotion'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'isNotInAlbum'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'isOffline'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'isReadOnly'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'isVisible'?: boolean;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'lensModel'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'libraryId'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'make'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'model'?: string;
+    /**
+     * 
+     * @type {AssetOrder}
+     * @memberof MetadataSearchDto
+     */
+    'order'?: AssetOrder;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'originalFileName'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'originalPath'?: string;
+    /**
+     * 
+     * @type {number}
+     * @memberof MetadataSearchDto
+     */
+    'page'?: number;
+    /**
+     * 
+     * @type {Array<string>}
+     * @memberof MetadataSearchDto
+     */
+    'personIds'?: Array<string>;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'resizePath'?: string;
+    /**
+     * 
+     * @type {number}
+     * @memberof MetadataSearchDto
+     */
+    'size'?: number;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'state'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'takenAfter'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'takenBefore'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'trashedAfter'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'trashedBefore'?: string;
+    /**
+     * 
+     * @type {AssetTypeEnum}
+     * @memberof MetadataSearchDto
+     */
+    'type'?: AssetTypeEnum;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'updatedAfter'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'updatedBefore'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof MetadataSearchDto
+     */
+    'webpPath'?: string;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'withArchived'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'withDeleted'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'withExif'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'withPeople'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof MetadataSearchDto
+     */
+    'withStacked'?: boolean;
+}
+
+
+/**
+ * 
+ * @export
  * @enum {string}
  */
 
@@ -3547,6 +3814,201 @@ export interface SmartInfoResponseDto {
      */
     'tags'?: Array<string> | null;
 }
+/**
+ * 
+ * @export
+ * @interface SmartSearchDto
+ */
+export interface SmartSearchDto {
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'city'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'country'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'createdAfter'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'createdBefore'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'deviceId'?: string;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'isArchived'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'isEncoded'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'isExternal'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'isFavorite'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'isMotion'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'isOffline'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'isReadOnly'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'isVisible'?: boolean;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'lensModel'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'libraryId'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'make'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'model'?: string;
+    /**
+     * 
+     * @type {number}
+     * @memberof SmartSearchDto
+     */
+    'page'?: number;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'query': string;
+    /**
+     * 
+     * @type {number}
+     * @memberof SmartSearchDto
+     */
+    'size'?: number;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'state'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'takenAfter'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'takenBefore'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'trashedAfter'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'trashedBefore'?: string;
+    /**
+     * 
+     * @type {AssetTypeEnum}
+     * @memberof SmartSearchDto
+     */
+    'type'?: AssetTypeEnum;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'updatedAfter'?: string;
+    /**
+     * 
+     * @type {string}
+     * @memberof SmartSearchDto
+     */
+    'updatedBefore'?: string;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'withArchived'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'withDeleted'?: boolean;
+    /**
+     * 
+     * @type {boolean}
+     * @memberof SmartSearchDto
+     */
+    'withExif'?: boolean;
+}
+
+
 /**
  * 
  * @export
@@ -7766,6 +8228,7 @@ export const AssetApiAxiosParamCreator = function (configuration?: Configuration
          * @param {boolean} [isExternal] 
          * @param {boolean} [isFavorite] 
          * @param {boolean} [isMotion] 
+         * @param {boolean} [isNotInAlbum] 
          * @param {boolean} [isOffline] 
          * @param {boolean} [isReadOnly] 
          * @param {boolean} [isVisible] 
@@ -7777,6 +8240,7 @@ export const AssetApiAxiosParamCreator = function (configuration?: Configuration
          * @param {string} [originalFileName] 
          * @param {string} [originalPath] 
          * @param {number} [page] 
+         * @param {Array<string>} [personIds] 
          * @param {string} [resizePath] 
          * @param {number} [size] 
          * @param {string} [state] 
@@ -7797,7 +8261,7 @@ export const AssetApiAxiosParamCreator = function (configuration?: Configuration
          * @deprecated
          * @throws {RequiredError}
          */
-        searchAssets: async (checksum?: string, city?: string, country?: string, createdAfter?: string, createdBefore?: string, deviceAssetId?: string, deviceId?: string, encodedVideoPath?: string, id?: string, isArchived?: boolean, isEncoded?: boolean, isExternal?: boolean, isFavorite?: boolean, isMotion?: boolean, isOffline?: boolean, isReadOnly?: boolean, isVisible?: boolean, lensModel?: string, libraryId?: string, make?: string, model?: string, order?: AssetOrder, originalFileName?: string, originalPath?: string, page?: number, resizePath?: string, size?: number, state?: string, takenAfter?: string, takenBefore?: string, trashedAfter?: string, trashedBefore?: string, type?: AssetTypeEnum, updatedAfter?: string, updatedBefore?: string, webpPath?: string, withArchived?: boolean, withDeleted?: boolean, withExif?: boolean, withPeople?: boolean, withStacked?: boolean, options: RawAxiosRequestConfig = {}): Promise<RequestArgs> => {
+        searchAssets: async (checksum?: string, city?: string, country?: string, createdAfter?: string, createdBefore?: string, deviceAssetId?: string, deviceId?: string, encodedVideoPath?: string, id?: string, isArchived?: boolean, isEncoded?: boolean, isExternal?: boolean, isFavorite?: boolean, isMotion?: boolean, isNotInAlbum?: boolean, isOffline?: boolean, isReadOnly?: boolean, isVisible?: boolean, lensModel?: string, libraryId?: string, make?: string, model?: string, order?: AssetOrder, originalFileName?: string, originalPath?: string, page?: number, personIds?: Array<string>, resizePath?: string, size?: number, state?: string, takenAfter?: string, takenBefore?: string, trashedAfter?: string, trashedBefore?: string, type?: AssetTypeEnum, updatedAfter?: string, updatedBefore?: string, webpPath?: string, withArchived?: boolean, withDeleted?: boolean, withExif?: boolean, withPeople?: boolean, withStacked?: boolean, options: RawAxiosRequestConfig = {}): Promise<RequestArgs> => {
             const localVarPath = `/assets`;
             // use dummy base URL string because the URL constructor only accepts absolute URLs.
             const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL);
@@ -7879,6 +8343,10 @@ export const AssetApiAxiosParamCreator = function (configuration?: Configuration
                 localVarQueryParameter['isMotion'] = isMotion;
             }
 
+            if (isNotInAlbum !== undefined) {
+                localVarQueryParameter['isNotInAlbum'] = isNotInAlbum;
+            }
+
             if (isOffline !== undefined) {
                 localVarQueryParameter['isOffline'] = isOffline;
             }
@@ -7921,6 +8389,10 @@ export const AssetApiAxiosParamCreator = function (configuration?: Configuration
 
             if (page !== undefined) {
                 localVarQueryParameter['page'] = page;
+            }
+
+            if (personIds) {
+                localVarQueryParameter['personIds'] = personIds;
             }
 
             if (resizePath !== undefined) {
@@ -8598,6 +9070,7 @@ export const AssetApiFp = function(configuration?: Configuration) {
          * @param {boolean} [isExternal] 
          * @param {boolean} [isFavorite] 
          * @param {boolean} [isMotion] 
+         * @param {boolean} [isNotInAlbum] 
          * @param {boolean} [isOffline] 
          * @param {boolean} [isReadOnly] 
          * @param {boolean} [isVisible] 
@@ -8609,6 +9082,7 @@ export const AssetApiFp = function(configuration?: Configuration) {
          * @param {string} [originalFileName] 
          * @param {string} [originalPath] 
          * @param {number} [page] 
+         * @param {Array<string>} [personIds] 
          * @param {string} [resizePath] 
          * @param {number} [size] 
          * @param {string} [state] 
@@ -8629,8 +9103,8 @@ export const AssetApiFp = function(configuration?: Configuration) {
          * @deprecated
          * @throws {RequiredError}
          */
-        async searchAssets(checksum?: string, city?: string, country?: string, createdAfter?: string, createdBefore?: string, deviceAssetId?: string, deviceId?: string, encodedVideoPath?: string, id?: string, isArchived?: boolean, isEncoded?: boolean, isExternal?: boolean, isFavorite?: boolean, isMotion?: boolean, isOffline?: boolean, isReadOnly?: boolean, isVisible?: boolean, lensModel?: string, libraryId?: string, make?: string, model?: string, order?: AssetOrder, originalFileName?: string, originalPath?: string, page?: number, resizePath?: string, size?: number, state?: string, takenAfter?: string, takenBefore?: string, trashedAfter?: string, trashedBefore?: string, type?: AssetTypeEnum, updatedAfter?: string, updatedBefore?: string, webpPath?: string, withArchived?: boolean, withDeleted?: boolean, withExif?: boolean, withPeople?: boolean, withStacked?: boolean, options?: RawAxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<Array<AssetResponseDto>>> {
-            const localVarAxiosArgs = await localVarAxiosParamCreator.searchAssets(checksum, city, country, createdAfter, createdBefore, deviceAssetId, deviceId, encodedVideoPath, id, isArchived, isEncoded, isExternal, isFavorite, isMotion, isOffline, isReadOnly, isVisible, lensModel, libraryId, make, model, order, originalFileName, originalPath, page, resizePath, size, state, takenAfter, takenBefore, trashedAfter, trashedBefore, type, updatedAfter, updatedBefore, webpPath, withArchived, withDeleted, withExif, withPeople, withStacked, options);
+        async searchAssets(checksum?: string, city?: string, country?: string, createdAfter?: string, createdBefore?: string, deviceAssetId?: string, deviceId?: string, encodedVideoPath?: string, id?: string, isArchived?: boolean, isEncoded?: boolean, isExternal?: boolean, isFavorite?: boolean, isMotion?: boolean, isNotInAlbum?: boolean, isOffline?: boolean, isReadOnly?: boolean, isVisible?: boolean, lensModel?: string, libraryId?: string, make?: string, model?: string, order?: AssetOrder, originalFileName?: string, originalPath?: string, page?: number, personIds?: Array<string>, resizePath?: string, size?: number, state?: string, takenAfter?: string, takenBefore?: string, trashedAfter?: string, trashedBefore?: string, type?: AssetTypeEnum, updatedAfter?: string, updatedBefore?: string, webpPath?: string, withArchived?: boolean, withDeleted?: boolean, withExif?: boolean, withPeople?: boolean, withStacked?: boolean, options?: RawAxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<Array<AssetResponseDto>>> {
+            const localVarAxiosArgs = await localVarAxiosParamCreator.searchAssets(checksum, city, country, createdAfter, createdBefore, deviceAssetId, deviceId, encodedVideoPath, id, isArchived, isEncoded, isExternal, isFavorite, isMotion, isNotInAlbum, isOffline, isReadOnly, isVisible, lensModel, libraryId, make, model, order, originalFileName, originalPath, page, personIds, resizePath, size, state, takenAfter, takenBefore, trashedAfter, trashedBefore, type, updatedAfter, updatedBefore, webpPath, withArchived, withDeleted, withExif, withPeople, withStacked, options);
             const index = configuration?.serverIndex ?? 0;
             const operationBasePath = operationServerMap['AssetApi.searchAssets']?.[index]?.url;
             return (axios, basePath) => createRequestFunction(localVarAxiosArgs, globalAxios, BASE_PATH, configuration)(axios, operationBasePath || basePath);
@@ -8882,7 +9356,7 @@ export const AssetApiFactory = function (configuration?: Configuration, basePath
          * @throws {RequiredError}
          */
         searchAssets(requestParameters: AssetApiSearchAssetsRequest = {}, options?: RawAxiosRequestConfig): AxiosPromise<Array<AssetResponseDto>> {
-            return localVarFp.searchAssets(requestParameters.checksum, requestParameters.city, requestParameters.country, requestParameters.createdAfter, requestParameters.createdBefore, requestParameters.deviceAssetId, requestParameters.deviceId, requestParameters.encodedVideoPath, requestParameters.id, requestParameters.isArchived, requestParameters.isEncoded, requestParameters.isExternal, requestParameters.isFavorite, requestParameters.isMotion, requestParameters.isOffline, requestParameters.isReadOnly, requestParameters.isVisible, requestParameters.lensModel, requestParameters.libraryId, requestParameters.make, requestParameters.model, requestParameters.order, requestParameters.originalFileName, requestParameters.originalPath, requestParameters.page, requestParameters.resizePath, requestParameters.size, requestParameters.state, requestParameters.takenAfter, requestParameters.takenBefore, requestParameters.trashedAfter, requestParameters.trashedBefore, requestParameters.type, requestParameters.updatedAfter, requestParameters.updatedBefore, requestParameters.webpPath, requestParameters.withArchived, requestParameters.withDeleted, requestParameters.withExif, requestParameters.withPeople, requestParameters.withStacked, options).then((request) => request(axios, basePath));
+            return localVarFp.searchAssets(requestParameters.checksum, requestParameters.city, requestParameters.country, requestParameters.createdAfter, requestParameters.createdBefore, requestParameters.deviceAssetId, requestParameters.deviceId, requestParameters.encodedVideoPath, requestParameters.id, requestParameters.isArchived, requestParameters.isEncoded, requestParameters.isExternal, requestParameters.isFavorite, requestParameters.isMotion, requestParameters.isNotInAlbum, requestParameters.isOffline, requestParameters.isReadOnly, requestParameters.isVisible, requestParameters.lensModel, requestParameters.libraryId, requestParameters.make, requestParameters.model, requestParameters.order, requestParameters.originalFileName, requestParameters.originalPath, requestParameters.page, requestParameters.personIds, requestParameters.resizePath, requestParameters.size, requestParameters.state, requestParameters.takenAfter, requestParameters.takenBefore, requestParameters.trashedAfter, requestParameters.trashedBefore, requestParameters.type, requestParameters.updatedAfter, requestParameters.updatedBefore, requestParameters.webpPath, requestParameters.withArchived, requestParameters.withDeleted, requestParameters.withExif, requestParameters.withPeople, requestParameters.withStacked, options).then((request) => request(axios, basePath));
         },
         /**
          * 
@@ -9489,6 +9963,13 @@ export interface AssetApiSearchAssetsRequest {
      * @type {boolean}
      * @memberof AssetApiSearchAssets
      */
+    readonly isNotInAlbum?: boolean
+
+    /**
+     * 
+     * @type {boolean}
+     * @memberof AssetApiSearchAssets
+     */
     readonly isOffline?: boolean
 
     /**
@@ -9560,6 +10041,13 @@ export interface AssetApiSearchAssetsRequest {
      * @memberof AssetApiSearchAssets
      */
     readonly page?: number
+
+    /**
+     * 
+     * @type {Array<string>}
+     * @memberof AssetApiSearchAssets
+     */
+    readonly personIds?: Array<string>
 
     /**
      * 
@@ -10077,7 +10565,7 @@ export class AssetApi extends BaseAPI {
      * @memberof AssetApi
      */
     public searchAssets(requestParameters: AssetApiSearchAssetsRequest = {}, options?: RawAxiosRequestConfig) {
-        return AssetApiFp(this.configuration).searchAssets(requestParameters.checksum, requestParameters.city, requestParameters.country, requestParameters.createdAfter, requestParameters.createdBefore, requestParameters.deviceAssetId, requestParameters.deviceId, requestParameters.encodedVideoPath, requestParameters.id, requestParameters.isArchived, requestParameters.isEncoded, requestParameters.isExternal, requestParameters.isFavorite, requestParameters.isMotion, requestParameters.isOffline, requestParameters.isReadOnly, requestParameters.isVisible, requestParameters.lensModel, requestParameters.libraryId, requestParameters.make, requestParameters.model, requestParameters.order, requestParameters.originalFileName, requestParameters.originalPath, requestParameters.page, requestParameters.resizePath, requestParameters.size, requestParameters.state, requestParameters.takenAfter, requestParameters.takenBefore, requestParameters.trashedAfter, requestParameters.trashedBefore, requestParameters.type, requestParameters.updatedAfter, requestParameters.updatedBefore, requestParameters.webpPath, requestParameters.withArchived, requestParameters.withDeleted, requestParameters.withExif, requestParameters.withPeople, requestParameters.withStacked, options).then((request) => request(this.axios, this.basePath));
+        return AssetApiFp(this.configuration).searchAssets(requestParameters.checksum, requestParameters.city, requestParameters.country, requestParameters.createdAfter, requestParameters.createdBefore, requestParameters.deviceAssetId, requestParameters.deviceId, requestParameters.encodedVideoPath, requestParameters.id, requestParameters.isArchived, requestParameters.isEncoded, requestParameters.isExternal, requestParameters.isFavorite, requestParameters.isMotion, requestParameters.isNotInAlbum, requestParameters.isOffline, requestParameters.isReadOnly, requestParameters.isVisible, requestParameters.lensModel, requestParameters.libraryId, requestParameters.make, requestParameters.model, requestParameters.order, requestParameters.originalFileName, requestParameters.originalPath, requestParameters.page, requestParameters.personIds, requestParameters.resizePath, requestParameters.size, requestParameters.state, requestParameters.takenAfter, requestParameters.takenBefore, requestParameters.trashedAfter, requestParameters.trashedBefore, requestParameters.type, requestParameters.updatedAfter, requestParameters.updatedBefore, requestParameters.webpPath, requestParameters.withArchived, requestParameters.withDeleted, requestParameters.withExif, requestParameters.withPeople, requestParameters.withStacked, options).then((request) => request(this.axios, this.basePath));
     }
 
     /**
@@ -14705,51 +15193,13 @@ export const SearchApiAxiosParamCreator = function (configuration?: Configuratio
         },
         /**
          * 
-         * @param {string} [checksum] 
-         * @param {string} [city] 
-         * @param {string} [country] 
-         * @param {string} [createdAfter] 
-         * @param {string} [createdBefore] 
-         * @param {string} [deviceAssetId] 
-         * @param {string} [deviceId] 
-         * @param {string} [encodedVideoPath] 
-         * @param {string} [id] 
-         * @param {boolean} [isArchived] 
-         * @param {boolean} [isEncoded] 
-         * @param {boolean} [isExternal] 
-         * @param {boolean} [isFavorite] 
-         * @param {boolean} [isMotion] 
-         * @param {boolean} [isOffline] 
-         * @param {boolean} [isReadOnly] 
-         * @param {boolean} [isVisible] 
-         * @param {string} [lensModel] 
-         * @param {string} [libraryId] 
-         * @param {string} [make] 
-         * @param {string} [model] 
-         * @param {AssetOrder} [order] 
-         * @param {string} [originalFileName] 
-         * @param {string} [originalPath] 
-         * @param {number} [page] 
-         * @param {string} [resizePath] 
-         * @param {number} [size] 
-         * @param {string} [state] 
-         * @param {string} [takenAfter] 
-         * @param {string} [takenBefore] 
-         * @param {string} [trashedAfter] 
-         * @param {string} [trashedBefore] 
-         * @param {AssetTypeEnum} [type] 
-         * @param {string} [updatedAfter] 
-         * @param {string} [updatedBefore] 
-         * @param {string} [webpPath] 
-         * @param {boolean} [withArchived] 
-         * @param {boolean} [withDeleted] 
-         * @param {boolean} [withExif] 
-         * @param {boolean} [withPeople] 
-         * @param {boolean} [withStacked] 
+         * @param {MetadataSearchDto} metadataSearchDto 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
-        searchMetadata: async (checksum?: string, city?: string, country?: string, createdAfter?: string, createdBefore?: string, deviceAssetId?: string, deviceId?: string, encodedVideoPath?: string, id?: string, isArchived?: boolean, isEncoded?: boolean, isExternal?: boolean, isFavorite?: boolean, isMotion?: boolean, isOffline?: boolean, isReadOnly?: boolean, isVisible?: boolean, lensModel?: string, libraryId?: string, make?: string, model?: string, order?: AssetOrder, originalFileName?: string, originalPath?: string, page?: number, resizePath?: string, size?: number, state?: string, takenAfter?: string, takenBefore?: string, trashedAfter?: string, trashedBefore?: string, type?: AssetTypeEnum, updatedAfter?: string, updatedBefore?: string, webpPath?: string, withArchived?: boolean, withDeleted?: boolean, withExif?: boolean, withPeople?: boolean, withStacked?: boolean, options: RawAxiosRequestConfig = {}): Promise<RequestArgs> => {
+        searchMetadata: async (metadataSearchDto: MetadataSearchDto, options: RawAxiosRequestConfig = {}): Promise<RequestArgs> => {
+            // verify required parameter 'metadataSearchDto' is not null or undefined
+            assertParamExists('searchMetadata', 'metadataSearchDto', metadataSearchDto)
             const localVarPath = `/search/metadata`;
             // use dummy base URL string because the URL constructor only accepts absolute URLs.
             const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL);
@@ -14758,7 +15208,7 @@ export const SearchApiAxiosParamCreator = function (configuration?: Configuratio
                 baseOptions = configuration.baseOptions;
             }
 
-            const localVarRequestOptions = { method: 'GET', ...baseOptions, ...options};
+            const localVarRequestOptions = { method: 'POST', ...baseOptions, ...options};
             const localVarHeaderParameter = {} as any;
             const localVarQueryParameter = {} as any;
 
@@ -14771,191 +15221,14 @@ export const SearchApiAxiosParamCreator = function (configuration?: Configuratio
             // http bearer authentication required
             await setBearerAuthToObject(localVarHeaderParameter, configuration)
 
-            if (checksum !== undefined) {
-                localVarQueryParameter['checksum'] = checksum;
-            }
-
-            if (city !== undefined) {
-                localVarQueryParameter['city'] = city;
-            }
-
-            if (country !== undefined) {
-                localVarQueryParameter['country'] = country;
-            }
-
-            if (createdAfter !== undefined) {
-                localVarQueryParameter['createdAfter'] = (createdAfter as any instanceof Date) ?
-                    (createdAfter as any).toISOString() :
-                    createdAfter;
-            }
-
-            if (createdBefore !== undefined) {
-                localVarQueryParameter['createdBefore'] = (createdBefore as any instanceof Date) ?
-                    (createdBefore as any).toISOString() :
-                    createdBefore;
-            }
-
-            if (deviceAssetId !== undefined) {
-                localVarQueryParameter['deviceAssetId'] = deviceAssetId;
-            }
-
-            if (deviceId !== undefined) {
-                localVarQueryParameter['deviceId'] = deviceId;
-            }
-
-            if (encodedVideoPath !== undefined) {
-                localVarQueryParameter['encodedVideoPath'] = encodedVideoPath;
-            }
-
-            if (id !== undefined) {
-                localVarQueryParameter['id'] = id;
-            }
-
-            if (isArchived !== undefined) {
-                localVarQueryParameter['isArchived'] = isArchived;
-            }
-
-            if (isEncoded !== undefined) {
-                localVarQueryParameter['isEncoded'] = isEncoded;
-            }
-
-            if (isExternal !== undefined) {
-                localVarQueryParameter['isExternal'] = isExternal;
-            }
-
-            if (isFavorite !== undefined) {
-                localVarQueryParameter['isFavorite'] = isFavorite;
-            }
-
-            if (isMotion !== undefined) {
-                localVarQueryParameter['isMotion'] = isMotion;
-            }
-
-            if (isOffline !== undefined) {
-                localVarQueryParameter['isOffline'] = isOffline;
-            }
-
-            if (isReadOnly !== undefined) {
-                localVarQueryParameter['isReadOnly'] = isReadOnly;
-            }
-
-            if (isVisible !== undefined) {
-                localVarQueryParameter['isVisible'] = isVisible;
-            }
-
-            if (lensModel !== undefined) {
-                localVarQueryParameter['lensModel'] = lensModel;
-            }
-
-            if (libraryId !== undefined) {
-                localVarQueryParameter['libraryId'] = libraryId;
-            }
-
-            if (make !== undefined) {
-                localVarQueryParameter['make'] = make;
-            }
-
-            if (model !== undefined) {
-                localVarQueryParameter['model'] = model;
-            }
-
-            if (order !== undefined) {
-                localVarQueryParameter['order'] = order;
-            }
-
-            if (originalFileName !== undefined) {
-                localVarQueryParameter['originalFileName'] = originalFileName;
-            }
-
-            if (originalPath !== undefined) {
-                localVarQueryParameter['originalPath'] = originalPath;
-            }
-
-            if (page !== undefined) {
-                localVarQueryParameter['page'] = page;
-            }
-
-            if (resizePath !== undefined) {
-                localVarQueryParameter['resizePath'] = resizePath;
-            }
-
-            if (size !== undefined) {
-                localVarQueryParameter['size'] = size;
-            }
-
-            if (state !== undefined) {
-                localVarQueryParameter['state'] = state;
-            }
-
-            if (takenAfter !== undefined) {
-                localVarQueryParameter['takenAfter'] = (takenAfter as any instanceof Date) ?
-                    (takenAfter as any).toISOString() :
-                    takenAfter;
-            }
-
-            if (takenBefore !== undefined) {
-                localVarQueryParameter['takenBefore'] = (takenBefore as any instanceof Date) ?
-                    (takenBefore as any).toISOString() :
-                    takenBefore;
-            }
-
-            if (trashedAfter !== undefined) {
-                localVarQueryParameter['trashedAfter'] = (trashedAfter as any instanceof Date) ?
-                    (trashedAfter as any).toISOString() :
-                    trashedAfter;
-            }
-
-            if (trashedBefore !== undefined) {
-                localVarQueryParameter['trashedBefore'] = (trashedBefore as any instanceof Date) ?
-                    (trashedBefore as any).toISOString() :
-                    trashedBefore;
-            }
-
-            if (type !== undefined) {
-                localVarQueryParameter['type'] = type;
-            }
-
-            if (updatedAfter !== undefined) {
-                localVarQueryParameter['updatedAfter'] = (updatedAfter as any instanceof Date) ?
-                    (updatedAfter as any).toISOString() :
-                    updatedAfter;
-            }
-
-            if (updatedBefore !== undefined) {
-                localVarQueryParameter['updatedBefore'] = (updatedBefore as any instanceof Date) ?
-                    (updatedBefore as any).toISOString() :
-                    updatedBefore;
-            }
-
-            if (webpPath !== undefined) {
-                localVarQueryParameter['webpPath'] = webpPath;
-            }
-
-            if (withArchived !== undefined) {
-                localVarQueryParameter['withArchived'] = withArchived;
-            }
-
-            if (withDeleted !== undefined) {
-                localVarQueryParameter['withDeleted'] = withDeleted;
-            }
-
-            if (withExif !== undefined) {
-                localVarQueryParameter['withExif'] = withExif;
-            }
-
-            if (withPeople !== undefined) {
-                localVarQueryParameter['withPeople'] = withPeople;
-            }
-
-            if (withStacked !== undefined) {
-                localVarQueryParameter['withStacked'] = withStacked;
-            }
-
 
     
+            localVarHeaderParameter['Content-Type'] = 'application/json';
+
             setSearchParams(localVarUrlObj, localVarQueryParameter);
             let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
             localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
+            localVarRequestOptions.data = serializeDataIfNeeded(metadataSearchDto, localVarRequestOptions, configuration)
 
             return {
                 url: toPathString(localVarUrlObj),
@@ -15014,43 +15287,13 @@ export const SearchApiAxiosParamCreator = function (configuration?: Configuratio
         },
         /**
          * 
-         * @param {string} query 
-         * @param {string} [city] 
-         * @param {string} [country] 
-         * @param {string} [createdAfter] 
-         * @param {string} [createdBefore] 
-         * @param {string} [deviceId] 
-         * @param {boolean} [isArchived] 
-         * @param {boolean} [isEncoded] 
-         * @param {boolean} [isExternal] 
-         * @param {boolean} [isFavorite] 
-         * @param {boolean} [isMotion] 
-         * @param {boolean} [isOffline] 
-         * @param {boolean} [isReadOnly] 
-         * @param {boolean} [isVisible] 
-         * @param {string} [lensModel] 
-         * @param {string} [libraryId] 
-         * @param {string} [make] 
-         * @param {string} [model] 
-         * @param {number} [page] 
-         * @param {number} [size] 
-         * @param {string} [state] 
-         * @param {string} [takenAfter] 
-         * @param {string} [takenBefore] 
-         * @param {string} [trashedAfter] 
-         * @param {string} [trashedBefore] 
-         * @param {AssetTypeEnum} [type] 
-         * @param {string} [updatedAfter] 
-         * @param {string} [updatedBefore] 
-         * @param {boolean} [withArchived] 
-         * @param {boolean} [withDeleted] 
-         * @param {boolean} [withExif] 
+         * @param {SmartSearchDto} smartSearchDto 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
-        searchSmart: async (query: string, city?: string, country?: string, createdAfter?: string, createdBefore?: string, deviceId?: string, isArchived?: boolean, isEncoded?: boolean, isExternal?: boolean, isFavorite?: boolean, isMotion?: boolean, isOffline?: boolean, isReadOnly?: boolean, isVisible?: boolean, lensModel?: string, libraryId?: string, make?: string, model?: string, page?: number, size?: number, state?: string, takenAfter?: string, takenBefore?: string, trashedAfter?: string, trashedBefore?: string, type?: AssetTypeEnum, updatedAfter?: string, updatedBefore?: string, withArchived?: boolean, withDeleted?: boolean, withExif?: boolean, options: RawAxiosRequestConfig = {}): Promise<RequestArgs> => {
-            // verify required parameter 'query' is not null or undefined
-            assertParamExists('searchSmart', 'query', query)
+        searchSmart: async (smartSearchDto: SmartSearchDto, options: RawAxiosRequestConfig = {}): Promise<RequestArgs> => {
+            // verify required parameter 'smartSearchDto' is not null or undefined
+            assertParamExists('searchSmart', 'smartSearchDto', smartSearchDto)
             const localVarPath = `/search/smart`;
             // use dummy base URL string because the URL constructor only accepts absolute URLs.
             const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL);
@@ -15059,7 +15302,7 @@ export const SearchApiAxiosParamCreator = function (configuration?: Configuratio
                 baseOptions = configuration.baseOptions;
             }
 
-            const localVarRequestOptions = { method: 'GET', ...baseOptions, ...options};
+            const localVarRequestOptions = { method: 'POST', ...baseOptions, ...options};
             const localVarHeaderParameter = {} as any;
             const localVarQueryParameter = {} as any;
 
@@ -15072,151 +15315,14 @@ export const SearchApiAxiosParamCreator = function (configuration?: Configuratio
             // http bearer authentication required
             await setBearerAuthToObject(localVarHeaderParameter, configuration)
 
-            if (city !== undefined) {
-                localVarQueryParameter['city'] = city;
-            }
-
-            if (country !== undefined) {
-                localVarQueryParameter['country'] = country;
-            }
-
-            if (createdAfter !== undefined) {
-                localVarQueryParameter['createdAfter'] = (createdAfter as any instanceof Date) ?
-                    (createdAfter as any).toISOString() :
-                    createdAfter;
-            }
-
-            if (createdBefore !== undefined) {
-                localVarQueryParameter['createdBefore'] = (createdBefore as any instanceof Date) ?
-                    (createdBefore as any).toISOString() :
-                    createdBefore;
-            }
-
-            if (deviceId !== undefined) {
-                localVarQueryParameter['deviceId'] = deviceId;
-            }
-
-            if (isArchived !== undefined) {
-                localVarQueryParameter['isArchived'] = isArchived;
-            }
-
-            if (isEncoded !== undefined) {
-                localVarQueryParameter['isEncoded'] = isEncoded;
-            }
-
-            if (isExternal !== undefined) {
-                localVarQueryParameter['isExternal'] = isExternal;
-            }
-
-            if (isFavorite !== undefined) {
-                localVarQueryParameter['isFavorite'] = isFavorite;
-            }
-
-            if (isMotion !== undefined) {
-                localVarQueryParameter['isMotion'] = isMotion;
-            }
-
-            if (isOffline !== undefined) {
-                localVarQueryParameter['isOffline'] = isOffline;
-            }
-
-            if (isReadOnly !== undefined) {
-                localVarQueryParameter['isReadOnly'] = isReadOnly;
-            }
-
-            if (isVisible !== undefined) {
-                localVarQueryParameter['isVisible'] = isVisible;
-            }
-
-            if (lensModel !== undefined) {
-                localVarQueryParameter['lensModel'] = lensModel;
-            }
-
-            if (libraryId !== undefined) {
-                localVarQueryParameter['libraryId'] = libraryId;
-            }
-
-            if (make !== undefined) {
-                localVarQueryParameter['make'] = make;
-            }
-
-            if (model !== undefined) {
-                localVarQueryParameter['model'] = model;
-            }
-
-            if (page !== undefined) {
-                localVarQueryParameter['page'] = page;
-            }
-
-            if (query !== undefined) {
-                localVarQueryParameter['query'] = query;
-            }
-
-            if (size !== undefined) {
-                localVarQueryParameter['size'] = size;
-            }
-
-            if (state !== undefined) {
-                localVarQueryParameter['state'] = state;
-            }
-
-            if (takenAfter !== undefined) {
-                localVarQueryParameter['takenAfter'] = (takenAfter as any instanceof Date) ?
-                    (takenAfter as any).toISOString() :
-                    takenAfter;
-            }
-
-            if (takenBefore !== undefined) {
-                localVarQueryParameter['takenBefore'] = (takenBefore as any instanceof Date) ?
-                    (takenBefore as any).toISOString() :
-                    takenBefore;
-            }
-
-            if (trashedAfter !== undefined) {
-                localVarQueryParameter['trashedAfter'] = (trashedAfter as any instanceof Date) ?
-                    (trashedAfter as any).toISOString() :
-                    trashedAfter;
-            }
-
-            if (trashedBefore !== undefined) {
-                localVarQueryParameter['trashedBefore'] = (trashedBefore as any instanceof Date) ?
-                    (trashedBefore as any).toISOString() :
-                    trashedBefore;
-            }
-
-            if (type !== undefined) {
-                localVarQueryParameter['type'] = type;
-            }
-
-            if (updatedAfter !== undefined) {
-                localVarQueryParameter['updatedAfter'] = (updatedAfter as any instanceof Date) ?
-                    (updatedAfter as any).toISOString() :
-                    updatedAfter;
-            }
-
-            if (updatedBefore !== undefined) {
-                localVarQueryParameter['updatedBefore'] = (updatedBefore as any instanceof Date) ?
-                    (updatedBefore as any).toISOString() :
-                    updatedBefore;
-            }
-
-            if (withArchived !== undefined) {
-                localVarQueryParameter['withArchived'] = withArchived;
-            }
-
-            if (withDeleted !== undefined) {
-                localVarQueryParameter['withDeleted'] = withDeleted;
-            }
-
-            if (withExif !== undefined) {
-                localVarQueryParameter['withExif'] = withExif;
-            }
-
 
     
+            localVarHeaderParameter['Content-Type'] = 'application/json';
+
             setSearchParams(localVarUrlObj, localVarQueryParameter);
             let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
             localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
+            localVarRequestOptions.data = serializeDataIfNeeded(smartSearchDto, localVarRequestOptions, configuration)
 
             return {
                 url: toPathString(localVarUrlObj),
@@ -15284,52 +15390,12 @@ export const SearchApiFp = function(configuration?: Configuration) {
         },
         /**
          * 
-         * @param {string} [checksum] 
-         * @param {string} [city] 
-         * @param {string} [country] 
-         * @param {string} [createdAfter] 
-         * @param {string} [createdBefore] 
-         * @param {string} [deviceAssetId] 
-         * @param {string} [deviceId] 
-         * @param {string} [encodedVideoPath] 
-         * @param {string} [id] 
-         * @param {boolean} [isArchived] 
-         * @param {boolean} [isEncoded] 
-         * @param {boolean} [isExternal] 
-         * @param {boolean} [isFavorite] 
-         * @param {boolean} [isMotion] 
-         * @param {boolean} [isOffline] 
-         * @param {boolean} [isReadOnly] 
-         * @param {boolean} [isVisible] 
-         * @param {string} [lensModel] 
-         * @param {string} [libraryId] 
-         * @param {string} [make] 
-         * @param {string} [model] 
-         * @param {AssetOrder} [order] 
-         * @param {string} [originalFileName] 
-         * @param {string} [originalPath] 
-         * @param {number} [page] 
-         * @param {string} [resizePath] 
-         * @param {number} [size] 
-         * @param {string} [state] 
-         * @param {string} [takenAfter] 
-         * @param {string} [takenBefore] 
-         * @param {string} [trashedAfter] 
-         * @param {string} [trashedBefore] 
-         * @param {AssetTypeEnum} [type] 
-         * @param {string} [updatedAfter] 
-         * @param {string} [updatedBefore] 
-         * @param {string} [webpPath] 
-         * @param {boolean} [withArchived] 
-         * @param {boolean} [withDeleted] 
-         * @param {boolean} [withExif] 
-         * @param {boolean} [withPeople] 
-         * @param {boolean} [withStacked] 
+         * @param {MetadataSearchDto} metadataSearchDto 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
-        async searchMetadata(checksum?: string, city?: string, country?: string, createdAfter?: string, createdBefore?: string, deviceAssetId?: string, deviceId?: string, encodedVideoPath?: string, id?: string, isArchived?: boolean, isEncoded?: boolean, isExternal?: boolean, isFavorite?: boolean, isMotion?: boolean, isOffline?: boolean, isReadOnly?: boolean, isVisible?: boolean, lensModel?: string, libraryId?: string, make?: string, model?: string, order?: AssetOrder, originalFileName?: string, originalPath?: string, page?: number, resizePath?: string, size?: number, state?: string, takenAfter?: string, takenBefore?: string, trashedAfter?: string, trashedBefore?: string, type?: AssetTypeEnum, updatedAfter?: string, updatedBefore?: string, webpPath?: string, withArchived?: boolean, withDeleted?: boolean, withExif?: boolean, withPeople?: boolean, withStacked?: boolean, options?: RawAxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<SearchResponseDto>> {
-            const localVarAxiosArgs = await localVarAxiosParamCreator.searchMetadata(checksum, city, country, createdAfter, createdBefore, deviceAssetId, deviceId, encodedVideoPath, id, isArchived, isEncoded, isExternal, isFavorite, isMotion, isOffline, isReadOnly, isVisible, lensModel, libraryId, make, model, order, originalFileName, originalPath, page, resizePath, size, state, takenAfter, takenBefore, trashedAfter, trashedBefore, type, updatedAfter, updatedBefore, webpPath, withArchived, withDeleted, withExif, withPeople, withStacked, options);
+        async searchMetadata(metadataSearchDto: MetadataSearchDto, options?: RawAxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<SearchResponseDto>> {
+            const localVarAxiosArgs = await localVarAxiosParamCreator.searchMetadata(metadataSearchDto, options);
             const index = configuration?.serverIndex ?? 0;
             const operationBasePath = operationServerMap['SearchApi.searchMetadata']?.[index]?.url;
             return (axios, basePath) => createRequestFunction(localVarAxiosArgs, globalAxios, BASE_PATH, configuration)(axios, operationBasePath || basePath);
@@ -15349,42 +15415,12 @@ export const SearchApiFp = function(configuration?: Configuration) {
         },
         /**
          * 
-         * @param {string} query 
-         * @param {string} [city] 
-         * @param {string} [country] 
-         * @param {string} [createdAfter] 
-         * @param {string} [createdBefore] 
-         * @param {string} [deviceId] 
-         * @param {boolean} [isArchived] 
-         * @param {boolean} [isEncoded] 
-         * @param {boolean} [isExternal] 
-         * @param {boolean} [isFavorite] 
-         * @param {boolean} [isMotion] 
-         * @param {boolean} [isOffline] 
-         * @param {boolean} [isReadOnly] 
-         * @param {boolean} [isVisible] 
-         * @param {string} [lensModel] 
-         * @param {string} [libraryId] 
-         * @param {string} [make] 
-         * @param {string} [model] 
-         * @param {number} [page] 
-         * @param {number} [size] 
-         * @param {string} [state] 
-         * @param {string} [takenAfter] 
-         * @param {string} [takenBefore] 
-         * @param {string} [trashedAfter] 
-         * @param {string} [trashedBefore] 
-         * @param {AssetTypeEnum} [type] 
-         * @param {string} [updatedAfter] 
-         * @param {string} [updatedBefore] 
-         * @param {boolean} [withArchived] 
-         * @param {boolean} [withDeleted] 
-         * @param {boolean} [withExif] 
+         * @param {SmartSearchDto} smartSearchDto 
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
-        async searchSmart(query: string, city?: string, country?: string, createdAfter?: string, createdBefore?: string, deviceId?: string, isArchived?: boolean, isEncoded?: boolean, isExternal?: boolean, isFavorite?: boolean, isMotion?: boolean, isOffline?: boolean, isReadOnly?: boolean, isVisible?: boolean, lensModel?: string, libraryId?: string, make?: string, model?: string, page?: number, size?: number, state?: string, takenAfter?: string, takenBefore?: string, trashedAfter?: string, trashedBefore?: string, type?: AssetTypeEnum, updatedAfter?: string, updatedBefore?: string, withArchived?: boolean, withDeleted?: boolean, withExif?: boolean, options?: RawAxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<SearchResponseDto>> {
-            const localVarAxiosArgs = await localVarAxiosParamCreator.searchSmart(query, city, country, createdAfter, createdBefore, deviceId, isArchived, isEncoded, isExternal, isFavorite, isMotion, isOffline, isReadOnly, isVisible, lensModel, libraryId, make, model, page, size, state, takenAfter, takenBefore, trashedAfter, trashedBefore, type, updatedAfter, updatedBefore, withArchived, withDeleted, withExif, options);
+        async searchSmart(smartSearchDto: SmartSearchDto, options?: RawAxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<SearchResponseDto>> {
+            const localVarAxiosArgs = await localVarAxiosParamCreator.searchSmart(smartSearchDto, options);
             const index = configuration?.serverIndex ?? 0;
             const operationBasePath = operationServerMap['SearchApi.searchSmart']?.[index]?.url;
             return (axios, basePath) => createRequestFunction(localVarAxiosArgs, globalAxios, BASE_PATH, configuration)(axios, operationBasePath || basePath);
@@ -15432,8 +15468,8 @@ export const SearchApiFactory = function (configuration?: Configuration, basePat
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
-        searchMetadata(requestParameters: SearchApiSearchMetadataRequest = {}, options?: RawAxiosRequestConfig): AxiosPromise<SearchResponseDto> {
-            return localVarFp.searchMetadata(requestParameters.checksum, requestParameters.city, requestParameters.country, requestParameters.createdAfter, requestParameters.createdBefore, requestParameters.deviceAssetId, requestParameters.deviceId, requestParameters.encodedVideoPath, requestParameters.id, requestParameters.isArchived, requestParameters.isEncoded, requestParameters.isExternal, requestParameters.isFavorite, requestParameters.isMotion, requestParameters.isOffline, requestParameters.isReadOnly, requestParameters.isVisible, requestParameters.lensModel, requestParameters.libraryId, requestParameters.make, requestParameters.model, requestParameters.order, requestParameters.originalFileName, requestParameters.originalPath, requestParameters.page, requestParameters.resizePath, requestParameters.size, requestParameters.state, requestParameters.takenAfter, requestParameters.takenBefore, requestParameters.trashedAfter, requestParameters.trashedBefore, requestParameters.type, requestParameters.updatedAfter, requestParameters.updatedBefore, requestParameters.webpPath, requestParameters.withArchived, requestParameters.withDeleted, requestParameters.withExif, requestParameters.withPeople, requestParameters.withStacked, options).then((request) => request(axios, basePath));
+        searchMetadata(requestParameters: SearchApiSearchMetadataRequest, options?: RawAxiosRequestConfig): AxiosPromise<SearchResponseDto> {
+            return localVarFp.searchMetadata(requestParameters.metadataSearchDto, options).then((request) => request(axios, basePath));
         },
         /**
          * 
@@ -15451,7 +15487,7 @@ export const SearchApiFactory = function (configuration?: Configuration, basePat
          * @throws {RequiredError}
          */
         searchSmart(requestParameters: SearchApiSearchSmartRequest, options?: RawAxiosRequestConfig): AxiosPromise<SearchResponseDto> {
-            return localVarFp.searchSmart(requestParameters.query, requestParameters.city, requestParameters.country, requestParameters.createdAfter, requestParameters.createdBefore, requestParameters.deviceId, requestParameters.isArchived, requestParameters.isEncoded, requestParameters.isExternal, requestParameters.isFavorite, requestParameters.isMotion, requestParameters.isOffline, requestParameters.isReadOnly, requestParameters.isVisible, requestParameters.lensModel, requestParameters.libraryId, requestParameters.make, requestParameters.model, requestParameters.page, requestParameters.size, requestParameters.state, requestParameters.takenAfter, requestParameters.takenBefore, requestParameters.trashedAfter, requestParameters.trashedBefore, requestParameters.type, requestParameters.updatedAfter, requestParameters.updatedBefore, requestParameters.withArchived, requestParameters.withDeleted, requestParameters.withExif, options).then((request) => request(axios, basePath));
+            return localVarFp.searchSmart(requestParameters.smartSearchDto, options).then((request) => request(axios, basePath));
         },
     };
 };
@@ -15583,290 +15619,10 @@ export interface SearchApiSearchRequest {
 export interface SearchApiSearchMetadataRequest {
     /**
      * 
-     * @type {string}
+     * @type {MetadataSearchDto}
      * @memberof SearchApiSearchMetadata
      */
-    readonly checksum?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly city?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly country?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly createdAfter?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly createdBefore?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly deviceAssetId?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly deviceId?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly encodedVideoPath?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly id?: string
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly isArchived?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly isEncoded?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly isExternal?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly isFavorite?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly isMotion?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly isOffline?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly isReadOnly?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly isVisible?: boolean
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly lensModel?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly libraryId?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly make?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly model?: string
-
-    /**
-     * 
-     * @type {AssetOrder}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly order?: AssetOrder
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly originalFileName?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly originalPath?: string
-
-    /**
-     * 
-     * @type {number}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly page?: number
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly resizePath?: string
-
-    /**
-     * 
-     * @type {number}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly size?: number
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly state?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly takenAfter?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly takenBefore?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly trashedAfter?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly trashedBefore?: string
-
-    /**
-     * 
-     * @type {AssetTypeEnum}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly type?: AssetTypeEnum
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly updatedAfter?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly updatedBefore?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly webpPath?: string
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly withArchived?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly withDeleted?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly withExif?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly withPeople?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchMetadata
-     */
-    readonly withStacked?: boolean
+    readonly metadataSearchDto: MetadataSearchDto
 }
 
 /**
@@ -15898,220 +15654,10 @@ export interface SearchApiSearchPersonRequest {
 export interface SearchApiSearchSmartRequest {
     /**
      * 
-     * @type {string}
+     * @type {SmartSearchDto}
      * @memberof SearchApiSearchSmart
      */
-    readonly query: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly city?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly country?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly createdAfter?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly createdBefore?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly deviceId?: string
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly isArchived?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly isEncoded?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly isExternal?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly isFavorite?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly isMotion?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly isOffline?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly isReadOnly?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly isVisible?: boolean
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly lensModel?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly libraryId?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly make?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly model?: string
-
-    /**
-     * 
-     * @type {number}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly page?: number
-
-    /**
-     * 
-     * @type {number}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly size?: number
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly state?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly takenAfter?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly takenBefore?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly trashedAfter?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly trashedBefore?: string
-
-    /**
-     * 
-     * @type {AssetTypeEnum}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly type?: AssetTypeEnum
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly updatedAfter?: string
-
-    /**
-     * 
-     * @type {string}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly updatedBefore?: string
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly withArchived?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly withDeleted?: boolean
-
-    /**
-     * 
-     * @type {boolean}
-     * @memberof SearchApiSearchSmart
-     */
-    readonly withExif?: boolean
+    readonly smartSearchDto: SmartSearchDto
 }
 
 /**
@@ -16161,8 +15707,8 @@ export class SearchApi extends BaseAPI {
      * @throws {RequiredError}
      * @memberof SearchApi
      */
-    public searchMetadata(requestParameters: SearchApiSearchMetadataRequest = {}, options?: RawAxiosRequestConfig) {
-        return SearchApiFp(this.configuration).searchMetadata(requestParameters.checksum, requestParameters.city, requestParameters.country, requestParameters.createdAfter, requestParameters.createdBefore, requestParameters.deviceAssetId, requestParameters.deviceId, requestParameters.encodedVideoPath, requestParameters.id, requestParameters.isArchived, requestParameters.isEncoded, requestParameters.isExternal, requestParameters.isFavorite, requestParameters.isMotion, requestParameters.isOffline, requestParameters.isReadOnly, requestParameters.isVisible, requestParameters.lensModel, requestParameters.libraryId, requestParameters.make, requestParameters.model, requestParameters.order, requestParameters.originalFileName, requestParameters.originalPath, requestParameters.page, requestParameters.resizePath, requestParameters.size, requestParameters.state, requestParameters.takenAfter, requestParameters.takenBefore, requestParameters.trashedAfter, requestParameters.trashedBefore, requestParameters.type, requestParameters.updatedAfter, requestParameters.updatedBefore, requestParameters.webpPath, requestParameters.withArchived, requestParameters.withDeleted, requestParameters.withExif, requestParameters.withPeople, requestParameters.withStacked, options).then((request) => request(this.axios, this.basePath));
+    public searchMetadata(requestParameters: SearchApiSearchMetadataRequest, options?: RawAxiosRequestConfig) {
+        return SearchApiFp(this.configuration).searchMetadata(requestParameters.metadataSearchDto, options).then((request) => request(this.axios, this.basePath));
     }
 
     /**
@@ -16184,7 +15730,7 @@ export class SearchApi extends BaseAPI {
      * @memberof SearchApi
      */
     public searchSmart(requestParameters: SearchApiSearchSmartRequest, options?: RawAxiosRequestConfig) {
-        return SearchApiFp(this.configuration).searchSmart(requestParameters.query, requestParameters.city, requestParameters.country, requestParameters.createdAfter, requestParameters.createdBefore, requestParameters.deviceId, requestParameters.isArchived, requestParameters.isEncoded, requestParameters.isExternal, requestParameters.isFavorite, requestParameters.isMotion, requestParameters.isOffline, requestParameters.isReadOnly, requestParameters.isVisible, requestParameters.lensModel, requestParameters.libraryId, requestParameters.make, requestParameters.model, requestParameters.page, requestParameters.size, requestParameters.state, requestParameters.takenAfter, requestParameters.takenBefore, requestParameters.trashedAfter, requestParameters.trashedBefore, requestParameters.type, requestParameters.updatedAfter, requestParameters.updatedBefore, requestParameters.withArchived, requestParameters.withDeleted, requestParameters.withExif, options).then((request) => request(this.axios, this.basePath));
+        return SearchApiFp(this.configuration).searchSmart(requestParameters.smartSearchDto, options).then((request) => request(this.axios, this.basePath));
     }
 }
 
