@@ -588,6 +588,84 @@ export type SearchExploreResponseDto = {
     fieldName: string;
     items: SearchExploreItem[];
 };
+export type MetadataSearchDto = {
+    checksum?: string;
+    city?: string;
+    country?: string;
+    createdAfter?: string;
+    createdBefore?: string;
+    deviceAssetId?: string;
+    deviceId?: string;
+    encodedVideoPath?: string;
+    id?: string;
+    isArchived?: boolean;
+    isEncoded?: boolean;
+    isExternal?: boolean;
+    isFavorite?: boolean;
+    isMotion?: boolean;
+    isNotInAlbum?: boolean;
+    isOffline?: boolean;
+    isReadOnly?: boolean;
+    isVisible?: boolean;
+    lensModel?: string;
+    libraryId?: string;
+    make?: string;
+    model?: string;
+    order?: AssetOrder;
+    originalFileName?: string;
+    originalPath?: string;
+    page?: number;
+    personIds?: string[];
+    resizePath?: string;
+    size?: number;
+    state?: string;
+    takenAfter?: string;
+    takenBefore?: string;
+    trashedAfter?: string;
+    trashedBefore?: string;
+    "type"?: AssetTypeEnum;
+    updatedAfter?: string;
+    updatedBefore?: string;
+    webpPath?: string;
+    withArchived?: boolean;
+    withDeleted?: boolean;
+    withExif?: boolean;
+    withPeople?: boolean;
+    withStacked?: boolean;
+};
+export type SmartSearchDto = {
+    city?: string;
+    country?: string;
+    createdAfter?: string;
+    createdBefore?: string;
+    deviceId?: string;
+    isArchived?: boolean;
+    isEncoded?: boolean;
+    isExternal?: boolean;
+    isFavorite?: boolean;
+    isMotion?: boolean;
+    isOffline?: boolean;
+    isReadOnly?: boolean;
+    isVisible?: boolean;
+    lensModel?: string;
+    libraryId?: string;
+    make?: string;
+    model?: string;
+    page?: number;
+    query: string;
+    size?: number;
+    state?: string;
+    takenAfter?: string;
+    takenBefore?: string;
+    trashedAfter?: string;
+    trashedBefore?: string;
+    "type"?: AssetTypeEnum;
+    updatedAfter?: string;
+    updatedBefore?: string;
+    withArchived?: boolean;
+    withDeleted?: boolean;
+    withExif?: boolean;
+};
 export type ServerInfoResponseDto = {
     diskAvailable: string;
     diskAvailableRaw: number;
@@ -1437,7 +1515,7 @@ export function updateAsset({ id, updateAssetDto }: {
         body: updateAssetDto
     })));
 }
-export function searchAssets({ checksum, city, country, createdAfter, createdBefore, deviceAssetId, deviceId, encodedVideoPath, id, isArchived, isEncoded, isExternal, isFavorite, isMotion, isOffline, isReadOnly, isVisible, lensModel, libraryId, make, model, order, originalFileName, originalPath, page, resizePath, size, state, takenAfter, takenBefore, trashedAfter, trashedBefore, $type, updatedAfter, updatedBefore, webpPath, withArchived, withDeleted, withExif, withPeople, withStacked }: {
+export function searchAssets({ checksum, city, country, createdAfter, createdBefore, deviceAssetId, deviceId, encodedVideoPath, id, isArchived, isEncoded, isExternal, isFavorite, isMotion, isNotInAlbum, isOffline, isReadOnly, isVisible, lensModel, libraryId, make, model, order, originalFileName, originalPath, page, personIds, resizePath, size, state, takenAfter, takenBefore, trashedAfter, trashedBefore, $type, updatedAfter, updatedBefore, webpPath, withArchived, withDeleted, withExif, withPeople, withStacked }: {
     checksum?: string;
     city?: string;
     country?: string;
@@ -1452,6 +1530,7 @@ export function searchAssets({ checksum, city, country, createdAfter, createdBef
     isExternal?: boolean;
     isFavorite?: boolean;
     isMotion?: boolean;
+    isNotInAlbum?: boolean;
     isOffline?: boolean;
     isReadOnly?: boolean;
     isVisible?: boolean;
@@ -1463,6 +1542,7 @@ export function searchAssets({ checksum, city, country, createdAfter, createdBef
     originalFileName?: string;
     originalPath?: string;
     page?: number;
+    personIds?: string[];
     resizePath?: string;
     size?: number;
     state?: string;
@@ -1498,6 +1578,7 @@ export function searchAssets({ checksum, city, country, createdAfter, createdBef
         isExternal,
         isFavorite,
         isMotion,
+        isNotInAlbum,
         isOffline,
         isReadOnly,
         isVisible,
@@ -1509,6 +1590,7 @@ export function searchAssets({ checksum, city, country, createdAfter, createdBef
         originalFileName,
         originalPath,
         page,
+        personIds,
         resizePath,
         size,
         state,
@@ -2063,97 +2145,17 @@ export function getExploreData(opts?: Oazapfts.RequestOpts) {
         ...opts
     }));
 }
-export function searchMetadata({ checksum, city, country, createdAfter, createdBefore, deviceAssetId, deviceId, encodedVideoPath, id, isArchived, isEncoded, isExternal, isFavorite, isMotion, isOffline, isReadOnly, isVisible, lensModel, libraryId, make, model, order, originalFileName, originalPath, page, resizePath, size, state, takenAfter, takenBefore, trashedAfter, trashedBefore, $type, updatedAfter, updatedBefore, webpPath, withArchived, withDeleted, withExif, withPeople, withStacked }: {
-    checksum?: string;
-    city?: string;
-    country?: string;
-    createdAfter?: string;
-    createdBefore?: string;
-    deviceAssetId?: string;
-    deviceId?: string;
-    encodedVideoPath?: string;
-    id?: string;
-    isArchived?: boolean;
-    isEncoded?: boolean;
-    isExternal?: boolean;
-    isFavorite?: boolean;
-    isMotion?: boolean;
-    isOffline?: boolean;
-    isReadOnly?: boolean;
-    isVisible?: boolean;
-    lensModel?: string;
-    libraryId?: string;
-    make?: string;
-    model?: string;
-    order?: AssetOrder;
-    originalFileName?: string;
-    originalPath?: string;
-    page?: number;
-    resizePath?: string;
-    size?: number;
-    state?: string;
-    takenAfter?: string;
-    takenBefore?: string;
-    trashedAfter?: string;
-    trashedBefore?: string;
-    $type?: AssetTypeEnum;
-    updatedAfter?: string;
-    updatedBefore?: string;
-    webpPath?: string;
-    withArchived?: boolean;
-    withDeleted?: boolean;
-    withExif?: boolean;
-    withPeople?: boolean;
-    withStacked?: boolean;
+export function searchMetadata({ metadataSearchDto }: {
+    metadataSearchDto: MetadataSearchDto;
 }, opts?: Oazapfts.RequestOpts) {
     return oazapfts.ok(oazapfts.fetchJson<{
-        status: 200;
+        status: 201;
         data: SearchResponseDto;
-    }>(`/search/metadata${QS.query(QS.explode({
-        checksum,
-        city,
-        country,
-        createdAfter,
-        createdBefore,
-        deviceAssetId,
-        deviceId,
-        encodedVideoPath,
-        id,
-        isArchived,
-        isEncoded,
-        isExternal,
-        isFavorite,
-        isMotion,
-        isOffline,
-        isReadOnly,
-        isVisible,
-        lensModel,
-        libraryId,
-        make,
-        model,
-        order,
-        originalFileName,
-        originalPath,
-        page,
-        resizePath,
-        size,
-        state,
-        takenAfter,
-        takenBefore,
-        trashedAfter,
-        trashedBefore,
-        "type": $type,
-        updatedAfter,
-        updatedBefore,
-        webpPath,
-        withArchived,
-        withDeleted,
-        withExif,
-        withPeople,
-        withStacked
-    }))}`, {
-        ...opts
-    }));
+    }>("/search/metadata", oazapfts.json({
+        ...opts,
+        method: "POST",
+        body: metadataSearchDto
+    })));
 }
 export function searchPerson({ name, withHidden }: {
     name: string;
@@ -2169,77 +2171,17 @@ export function searchPerson({ name, withHidden }: {
         ...opts
     }));
 }
-export function searchSmart({ city, country, createdAfter, createdBefore, deviceId, isArchived, isEncoded, isExternal, isFavorite, isMotion, isOffline, isReadOnly, isVisible, lensModel, libraryId, make, model, page, query, size, state, takenAfter, takenBefore, trashedAfter, trashedBefore, $type, updatedAfter, updatedBefore, withArchived, withDeleted, withExif }: {
-    city?: string;
-    country?: string;
-    createdAfter?: string;
-    createdBefore?: string;
-    deviceId?: string;
-    isArchived?: boolean;
-    isEncoded?: boolean;
-    isExternal?: boolean;
-    isFavorite?: boolean;
-    isMotion?: boolean;
-    isOffline?: boolean;
-    isReadOnly?: boolean;
-    isVisible?: boolean;
-    lensModel?: string;
-    libraryId?: string;
-    make?: string;
-    model?: string;
-    page?: number;
-    query: string;
-    size?: number;
-    state?: string;
-    takenAfter?: string;
-    takenBefore?: string;
-    trashedAfter?: string;
-    trashedBefore?: string;
-    $type?: AssetTypeEnum;
-    updatedAfter?: string;
-    updatedBefore?: string;
-    withArchived?: boolean;
-    withDeleted?: boolean;
-    withExif?: boolean;
+export function searchSmart({ smartSearchDto }: {
+    smartSearchDto: SmartSearchDto;
 }, opts?: Oazapfts.RequestOpts) {
     return oazapfts.ok(oazapfts.fetchJson<{
-        status: 200;
+        status: 201;
         data: SearchResponseDto;
-    }>(`/search/smart${QS.query(QS.explode({
-        city,
-        country,
-        createdAfter,
-        createdBefore,
-        deviceId,
-        isArchived,
-        isEncoded,
-        isExternal,
-        isFavorite,
-        isMotion,
-        isOffline,
-        isReadOnly,
-        isVisible,
-        lensModel,
-        libraryId,
-        make,
-        model,
-        page,
-        query,
-        size,
-        state,
-        takenAfter,
-        takenBefore,
-        trashedAfter,
-        trashedBefore,
-        "type": $type,
-        updatedAfter,
-        updatedBefore,
-        withArchived,
-        withDeleted,
-        withExif
-    }))}`, {
-        ...opts
-    }));
+    }>("/search/smart", oazapfts.json({
+        ...opts,
+        method: "POST",
+        body: smartSearchDto
+    })));
 }
 export function getSearchSuggestions({ country, make, model, state, $type }: {
     country?: string;
