@@ -6,13 +6,13 @@
 
   let showModal = false;
 
-  const { onRelease } = websocketStore;
+  const { release } = websocketStore;
 
   const semverToName = ({ major, minor, patch }: ServerVersionResponseDto) => `v${major}.${minor}.${patch}`;
 
-  $: releaseVersion = $onRelease && semverToName($onRelease.releaseVersion);
-  $: serverVersion = $onRelease && semverToName($onRelease.serverVersion);
-  $: $onRelease?.isAvailable && handleRelease();
+  $: releaseVersion = $release && semverToName($release.releaseVersion);
+  $: serverVersion = $release && semverToName($release.serverVersion);
+  $: $release?.isAvailable && handleRelease();
 
   const onAcknowledge = () => {
     localStorage.setItem('appVersion', releaseVersion);
